+++ packages/mantine/src/MantineValueEditor.tsx
@@ -38,13 +38,18 @@ export const MantineValueEditor = (allProps: MantineValueEditorProps): React.JSX
     selectorComponent: SelectorComponent = allProps.schema.controls.valueSelector,
     validation: _validation,
     extraProps,
-    inputType: _inputType,
+    inputType,
     parseNumbers: _parseNumbers,
     ...propsForValueSelector
   } = allProps;
 
-  const { valueAsArray, multiValueHandler, valueListItemClassName, inputTypeCoerced } =
-    useValueEditor(allProps);
+  const {
+    valueAsArray,
+    multiValueHandler,
+    valueListItemClassName,
+    inputTypeCoerced,
+    bigIntValueHandler,
+  } = useValueEditor(allProps);
 
   if (operator === 'null' || operator === 'notNull') {
     return null;
@@ -243,6 +248,22 @@ export const MantineValueEditor = (allProps: MantineValueEditorProps): React.JSX
           handleOnChange(d ? dayjs(d).format(dateFormat) : /* istanbul ignore next */ '')
         }
         popoverProps={{ withinPortal: false }}
+        {...extraProps}
+      />
+    );
+  }
+
+  if (inputType === 'bigint') {
+    return (
+      <TextInput
+        data-testid={testID}
+        type={inputTypeCoerced}
+        placeholder={placeHolderText}
+        value={`${value}`}
+        title={title}
+        className={className}
+        disabled={disabled}
+        onChange={e => bigIntValueHandler(e.target.value)}
         {...extraProps}
       />
     );