+++ packages/react-querybuilder/src/components/ValueEditor.tsx
@@ -21,6 +21,7 @@ export const ValueEditor = <F extends FullField>(
     title,
     className,
     type = 'text',
+    inputType,
     values = [],
     listsAsArrays,
     fieldData,
@@ -32,7 +33,6 @@ export const ValueEditor = <F extends FullField>(
     // we cherry pick these out of `propsForValueSelector` to keep them from being
     // assigned to DOM elements. (The props with mixed case are the only ones that
     // really matter. Props in all lowercase don't emit warnings.)
-    inputType: _inputType,
     parseNumbers: _parseNumbers,
     skipHook: _skipHook,
     valueSource: _valueSource,
@@ -42,6 +42,7 @@ export const ValueEditor = <F extends FullField>(
   const {
     valueAsArray,
     multiValueHandler,
+    bigIntValueHandler,
     parseNumberMethod,
     valueListItemClassName,
     inputTypeCoerced,
@@ -160,6 +161,23 @@ export const ValueEditor = <F extends FullField>(
       );
   }
 
+  // Note that we don't use `inputTypeCoerced` for this condition.
+  // We need to know that its uncoerced value is "bigint".
+  if (inputType === 'bigint') {
+    return (
+      <input
+        data-testid={testID}
+        type={inputTypeCoerced}
+        placeholder={placeHolderText}
+        value={`${value}`}
+        title={title}
+        className={className}
+        disabled={disabled}
+        onChange={e => bigIntValueHandler(e.target.value)}
+      />
+    );
+  }
+
   return (
     <input
       data-testid={testID}
@@ -175,6 +193,44 @@ export const ValueEditor = <F extends FullField>(
     />
   );
 };
+
+export interface UseValueEditor {
+  /**
+   * Array of values for when the main value represents a list, e.g. when operator
+   * is "between" or "in".
+   */
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  valueAsArray: any[];
+  /**
+   * An update handler for a series of value editors, e.g. when operator is "between".
+   * Calling this function will update a single element of the value array and leave
+   * the rest of the array as is.
+   *
+   * @param {string} val The new value for the editor
+   * @param {number} idx The index of the editor (and the array element to update)
+   */
+  multiValueHandler: (val: unknown, idx: number) => void;
+  /**
+   * An update handler for bigint editors, e.g. when `inputType` is "bigint" and
+   * `parseNumbersMethod` is truthy.
+   */
+  bigIntValueHandler: (val: unknown) => void;
+  /**
+   * Evaluated `parseNumber` method based on `parseNumbers` prop. This property ends up
+   * being the same as the `parseNumbers` prop minus the "-limited" suffix, unless
+   * the "-limited" suffix is present and the `inputType` is not "number", in which case
+   * it's set to `false`.
+   */
+  parseNumberMethod: ParseNumberMethod;
+  /**
+   * Class for items in a value editor series (e.g. "between" value editors).
+   */
+  valueListItemClassName: string;
+  /**
+   * Coerced `inputType` based on `inputType` and `operator`.
+   */
+  inputTypeCoerced: InputType;
+}
 
 /**
  * This hook is primarily concerned with multi-value editors like date range
@@ -210,39 +266,7 @@ export const ValueEditor = <F extends FullField>(
  */
 export const useValueEditor = <F extends FullField = FullField, O extends string = string>(
   props: ValueEditorProps<F, O>
-): {
-  /**
-   * Array of values for when the main value represents a list, e.g. when operator
-   * is "between" or "in".
-   */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  valueAsArray: any[];
-  /**
-   * An update handler for a series of value editors, e.g. when operator is "between".
-   * Calling this function will update a single element of the value array and leave
-   * the rest of the array as is.
-   *
-   * @param {string} val The new value for the editor
-   * @param {number} idx The index of the editor (and the array element to update)
-   */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  multiValueHandler: (val: any, idx: number) => void;
-  /**
-   * Evaluated `parseNumber` method based on `parseNumbers` prop. This property ends up
-   * being the same as the `parseNumbers` prop minus the "-limited" suffix, unless
-   * the "-limited" suffix is present and the `inputType` is not "number", in which case
-   * it's set to `false`.
-   */
-  parseNumberMethod: ParseNumberMethod;
-  /**
-   * Class for items in a value editor series (e.g. "between" value editors).
-   */
-  valueListItemClassName: string;
-  /**
-   * Coerced `inputType` based on `inputType` and `operator`.
-   */
-  inputTypeCoerced: InputType;
-} => {
+): UseValueEditor => {
   const {
     handleOnChange,
     inputType,
@@ -294,17 +318,39 @@ export const useValueEditor = <F extends FullField = FullField, O extends string
     [handleOnChange, listsAsArrays, operator, parseNumberMethod, valueAsArray, values]
   );
 
+  const bigIntValueHandler = useCallback(
+    (v: unknown) => {
+      const valAsMaybeNumber = parseNumber(v, {
+        parseNumbers: parseNumberMethod,
+        bigIntOnOverflow: true,
+      });
+      let bi: bigint;
+      try {
+        bi = BigInt(valAsMaybeNumber);
+      } catch {
+        handleOnChange(valAsMaybeNumber);
+        return;
+      }
+      handleOnChange(bi);
+    },
+    [handleOnChange, parseNumberMethod]
+  );
+
   const valueListItemClassName = clsx(
     suppressStandardClassnames || standardClassnames.valueListItem,
     // Optional chaining is necessary for QueryBuilderNative
     classNamesProp?.valueListItem
   );
 
-  const inputTypeCoerced = operator === 'in' || operator === 'notIn' ? 'text' : inputType || 'text';
+  const inputTypeCoerced =
+    inputType === 'bigint' || operator === 'in' || operator === 'notIn'
+      ? 'text'
+      : inputType || 'text';
 
   return {
     valueAsArray,
     multiValueHandler,
+    bigIntValueHandler,
     parseNumberMethod,
     valueListItemClassName,
     inputTypeCoerced,