+++ packages/tremor/src/TremorValueEditor.tsx
@@ -39,6 +39,7 @@ export const TremorValueEditor = (allProps: TremorValueEditorProps): React.JSX.E
     title,
     className,
     type,
+    inputType,
     values: _v,
     listsAsArrays,
     parseNumbers: _parseNumbers,
@@ -47,13 +48,17 @@ export const TremorValueEditor = (allProps: TremorValueEditorProps): React.JSX.E
     disabled,
     testID,
     selectorComponent = ClearableValueSelector,
-    inputType: _inputType,
     validation: _validation,
     extraProps,
   } = allProps;
 
-  const { valueAsArray, multiValueHandler, valueListItemClassName, inputTypeCoerced } =
-    useValueEditor(allProps);
+  const {
+    valueAsArray,
+    multiValueHandler,
+    bigIntValueHandler,
+    valueListItemClassName,
+    inputTypeCoerced,
+  } = useValueEditor(allProps);
 
   if (operator === 'null' || operator === 'notNull') {
     return null;
@@ -187,6 +192,22 @@ export const TremorValueEditor = (allProps: TremorValueEditorProps): React.JSX.E
         disabled={disabled}
         value={value}
         onValueChange={handleOnChange}
+        {...extraProps}
+      />
+    );
+  }
+
+  if (inputType === 'bigint') {
+    return (
+      <TextInput
+        data-testid={testID}
+        type={inputTypeCoerced as TextInputProps['type']}
+        placeholder={placeHolderText}
+        value={`${value}`}
+        title={title}
+        className={className}
+        disabled={disabled}
+        onChange={e => bigIntValueHandler(e.target.value)}
         {...extraProps}
       />
     );